+++ lib/impl/src/lib.rs
@@ -3,6 +3,8 @@ use arrow::ffi::FFI_ArrowSchema;
 
 #[no_mangle]
 pub extern "C" fn callwithschema(schema: &mut FFI_ArrowSchema)  {
-    println!("Hello from Rust, with FFI_ArrowSchema: {:?}", schema);
+//    println!("Hello from Rust, with FFI_ArrowSchema: {:?}", schema);
+    println!("Hello from Rust !");
+
 }
