+++ testapp.go
@@ -1,0 +1,21 @@
+package main
+
+import (
+	"fmt"
+	"github.com/apache/arrow/go/v9/arrow"
+	"github.com/apache/arrow/go/v9/arrow/memory"
+)
+
+func main() {
+	schema := arrow.NewSchema(
+		[]arrow.Field{
+			{Name: "f1-i32", Type: arrow.PrimitiveTypes.Int32},
+			{Name: "f2-f64", Type: arrow.PrimitiveTypes.Float64},
+		},
+		nil,
+	)
+
+	goBridge := GoBridge{GoAllocator: memory.NewGoAllocator()}
+	err := goBridge.Call(schema)
+	fmt.Println(err)
+}
